Make mockEntities exportable
